Move TcpFlag and related struct to separate module
